No longer runs on functions that aren't run
Previously we would run feature extraction on all functions regardless
of if they were run in our profile run. For example, running bzip2 in
profiling only runs compress, so the ~3.7M paths in uncompress are all
zero. Now we check if there are no real paths through a fx first.