Changed to correctly process unused paths
In an attempt to optimize I accidently removed a lot of information
that was present since the PathProfileInfo class IGNORES any path that
isn't run. This is obv not okay since we still want paths with zero
executions to have their features extracted. I had to revert to the old
code that does manual path iteration instead of using the class