Added benchmark lists, also added first classifier
This classifier does LOOCV over all of the polybench programs,
and applies logistic regression, SVM, and random forest models
to the data. I didn't really worry too much about hyperparameter
selection here, so these results can likely be improved. We can get
about a 0.90 AUC with logistic regression, and a 0.93 with random
forests. Not bad. Note SVMs perform poorly atm, although with a better
choice of hyperparameters and kernel choice I think we can definitely
match or beat random forests.

One important thing to remember about this script is that we have to be
able to fit the entire thing into memory for this to work out. That
means that for the SPEC programs we may need to change it up, or try
and get less samples somehow. Also note that because these are math
programs, there aren't that many paths and the class imbalance is
therefore small. Eventually the % positive examples will fall below an
acceptable level and we'll need to either oversample or otherwise
compensate.