LSTM stuff added
I've added an LSTM classifier as well as an LSTM feature extraction
pass. The feature extractor extracts the number of each opcode per basic
block and outputs these vectors for each function (per BB). This means
that the file has a short header with the basic block information,
followed by N lines of vectors per function, where N is num of basic
blocks. Check LSTMStaticEstimator for more information on this.

The LSTM itself is a two layer LSTM with 128 hidden units per layer.
This can change but I haven't been able to find a much better
architecture, and I've tried more layers and larger layers (as well as
GRU instead of LSTM). The LSTM takes in each vector from the basic block
and attempts to output the label at each step. The LSTM has a
mean-pooling output and then a softmax to try and predict the class. The
architecture is very similar to the Theano IMDB sentiment analysis
network.

For the numeric benchmarks, I do cross validation across each program
again. There isn't that much data, however, so I would think that the
larger programs will offer more advantage in LSTM. Even so, we can do
about as well as logistic regression but with just basic blocks.

Training will take forever if you don't have a GPU. We should probably
come up with a better method for this because LOOCV is hella slow for
deep learning models. Maybe just hold out a few benchmarks and train and
test accordingly. This would allow us to better view the train test
curves also.